Add metadata wrapper for T-shaped marketer tool
